Migrate form-related tests to shared infrastructure
