+++ src/Http/Http.Extensions/test/RequestDelegateFactoryTests.cs
@@ -2894,84 +2894,6 @@ public partial class RequestDelegateFactoryTests : LoggedTest
     }
 
     [Fact]
-    public async Task RequestDelegatePopulatesFromIFormFileCollectionParameter()
-    {
-        IFormFileCollection? formFilesArgument = null;
-
-        void TestAction(IFormFileCollection formFiles)
-        {
-            formFilesArgument = formFiles;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "file", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files, formFilesArgument);
-        Assert.NotNull(formFilesArgument!["file"]);
-
-        var allAcceptsMetadata = factoryResult.EndpointMetadata.OfType<IAcceptsMetadata>();
-        var acceptsMetadata = Assert.Single(allAcceptsMetadata);
-
-        Assert.NotNull(acceptsMetadata);
-        Assert.Equal(new[] { "multipart/form-data" }, acceptsMetadata.ContentTypes);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePopulatesFromIFormFileCollectionParameterWithAttribute()
-    {
-        IFormFileCollection? formFilesArgument = null;
-
-        void TestAction([FromForm] IFormFileCollection formFiles)
-        {
-            formFilesArgument = formFiles;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "file", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files, formFilesArgument);
-        Assert.NotNull(formFilesArgument!["file"]);
-
-        var allAcceptsMetadata = factoryResult.EndpointMetadata.OfType<IAcceptsMetadata>();
-        var acceptsMetadata = Assert.Single(allAcceptsMetadata);
-
-        Assert.NotNull(acceptsMetadata);
-        Assert.Equal(new[] { "multipart/form-data" }, acceptsMetadata.ContentTypes);
-    }
-
-    [Fact]
     public void CreateThrowsNotSupportedExceptionIfIFormFileCollectionHasMetadataParameterName()
     {
         IFormFileCollection? formFilesArgument = null;
@@ -2983,233 +2905,6 @@ public partial class RequestDelegateFactoryTests : LoggedTest
 
         var nse = Assert.Throws<NotSupportedException>(() => RequestDelegateFactory.Create(TestAction));
         Assert.Equal("Assigning a value to the IFromFormMetadata.Name property is not supported for parameters of type IFormFileCollection.", nse.Message);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePopulatesFromIFormFileParameter()
-    {
-        IFormFile? fileArgument = null;
-
-        void TestAction(IFormFile file)
-        {
-            fileArgument = file;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "file", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files["file"], fileArgument);
-        Assert.Equal("file.txt", fileArgument!.FileName);
-        Assert.Equal("file", fileArgument.Name);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePopulatesFromOptionalIFormFileParameter()
-    {
-        IFormFile? fileArgument = null;
-
-        void TestAction(IFormFile? file)
-        {
-            fileArgument = file;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "file", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files["file"], fileArgument);
-        Assert.Equal("file.txt", fileArgument!.FileName);
-        Assert.Equal("file", fileArgument.Name);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePopulatesFromMultipleRequiredIFormFileParameters()
-    {
-        IFormFile? file1Argument = null;
-        IFormFile? file2Argument = null;
-
-        void TestAction(IFormFile file1, IFormFile file2)
-        {
-            file1Argument = file1;
-            file2Argument = file2;
-        }
-
-        var fileContent1 = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var fileContent2 = new StringContent("there", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent1, "file1", "file1.txt");
-        form.Add(fileContent2, "file2", "file2.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files["file1"], file1Argument);
-        Assert.Equal("file1.txt", file1Argument!.FileName);
-        Assert.Equal("file1", file1Argument.Name);
-
-        Assert.Equal(httpContext.Request.Form.Files["file2"], file2Argument);
-        Assert.Equal("file2.txt", file2Argument!.FileName);
-        Assert.Equal("file2", file2Argument.Name);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePopulatesFromOptionalMissingIFormFileParameter()
-    {
-        IFormFile? file1Argument = null;
-        IFormFile? file2Argument = null;
-
-        void TestAction(IFormFile? file1, IFormFile? file2)
-        {
-            file1Argument = file1;
-            file2Argument = file2;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "file1", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files["file1"], file1Argument);
-        Assert.NotNull(file1Argument);
-
-        Assert.Equal(httpContext.Request.Form.Files["file2"], file2Argument);
-        Assert.Null(file2Argument);
-
-        var allAcceptsMetadata = factoryResult.EndpointMetadata.OfType<IAcceptsMetadata>();
-        var acceptsMetadata = Assert.Single(allAcceptsMetadata);
-
-        Assert.NotNull(acceptsMetadata);
-        Assert.Equal(new[] { "multipart/form-data" }, acceptsMetadata.ContentTypes);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePopulatesFromIFormFileParameterWithMetadata()
-    {
-        IFormFile? fileArgument = null;
-
-        void TestAction([FromForm(Name = "my_file")] IFormFile file)
-        {
-            fileArgument = file;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "my_file", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files["my_file"], fileArgument);
-        Assert.Equal("file.txt", fileArgument!.FileName);
-        Assert.Equal("my_file", fileArgument.Name);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePopulatesFromIFormFileAndBoundParameter()
-    {
-        IFormFile? fileArgument = null;
-        TraceIdentifier traceIdArgument = default;
-
-        void TestAction(IFormFile? file, TraceIdentifier traceId)
-        {
-            fileArgument = file;
-            traceIdArgument = traceId;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "file", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-        httpContext.TraceIdentifier = "my-trace-id";
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files["file"], fileArgument);
-        Assert.Equal("file.txt", fileArgument!.FileName);
-        Assert.Equal("file", fileArgument.Name);
-
-        Assert.Equal("my-trace-id", traceIdArgument.Id);
     }
 
     private readonly struct TraceIdentifier
@@ -3227,212 +2922,6 @@ public partial class RequestDelegateFactoryTests : LoggedTest
         {
             return ValueTask.FromResult(new TraceIdentifier(context.TraceIdentifier));
         }
-    }
-
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task RequestDelegateRejectsNonFormContent(bool shouldThrow)
-    {
-        var httpContext = new DefaultHttpContext();
-        httpContext.Request.Headers["Content-Type"] = "application/xml";
-        httpContext.Request.Headers["Content-Length"] = "1";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var serviceCollection = new ServiceCollection();
-        serviceCollection.AddSingleton(LoggerFactory);
-        httpContext.RequestServices = serviceCollection.BuildServiceProvider();
-
-        var factoryResult = RequestDelegateFactory.Create((HttpContext context, IFormFile file) =>
-        {
-        }, new RequestDelegateFactoryOptions() { ThrowOnBadRequest = shouldThrow });
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        var request = requestDelegate(httpContext);
-
-        if (shouldThrow)
-        {
-            var ex = await Assert.ThrowsAsync<BadHttpRequestException>(() => request);
-            Assert.Equal("Expected a supported form media type but got \"application/xml\".", ex.Message);
-            Assert.Equal(StatusCodes.Status415UnsupportedMediaType, ex.StatusCode);
-        }
-        else
-        {
-            await request;
-
-            Assert.Equal(415, httpContext.Response.StatusCode);
-            var logMessage = Assert.Single(TestSink.Writes);
-            Assert.Equal(new EventId(7, "UnexpectedContentType"), logMessage.EventId);
-            Assert.Equal(LogLevel.Debug, logMessage.LogLevel);
-            Assert.Equal("Expected a supported form media type but got \"application/xml\".", logMessage.Message);
-        }
-    }
-
-    [Fact]
-    public async Task RequestDelegateSets400ResponseIfRequiredFileNotSpecified()
-    {
-        var invoked = false;
-
-        void TestAction(IFormFile file)
-        {
-            invoked = true;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "some-other-file", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.False(invoked);
-        Assert.Equal(400, httpContext.Response.StatusCode);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePopulatesFromBothFormFileCollectionAndFormFileParameters()
-    {
-        IFormFileCollection? formFilesArgument = null;
-        IFormFile? fileArgument = null;
-
-        void TestAction(IFormFileCollection formFiles, IFormFile file)
-        {
-            formFilesArgument = formFiles;
-            fileArgument = file;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "file", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files, formFilesArgument);
-        Assert.NotNull(formFilesArgument!["file"]);
-
-        Assert.Equal(httpContext.Request.Form.Files["file"], fileArgument);
-        Assert.Equal("file.txt", fileArgument!.FileName);
-        Assert.Equal("file", fileArgument.Name);
-
-        var allAcceptsMetadata = factoryResult.EndpointMetadata.OfType<IAcceptsMetadata>();
-        var acceptsMetadata = Assert.Single(allAcceptsMetadata);
-
-        Assert.NotNull(acceptsMetadata);
-        Assert.Equal(new[] { "multipart/form-data" }, acceptsMetadata.ContentTypes);
-    }
-
-    [Theory]
-    [InlineData("Authorization", "bearer my-token")]
-    [InlineData("Cookie", ".AspNetCore.Auth=abc123")]
-    public async Task RequestDelegatePopulatesFromIFormFileParameterIfRequestContainsSecureHeader(
-        string headerName,
-        string headerValue)
-    {
-        IFormFile? fileArgument = null;
-        TraceIdentifier traceIdArgument = default;
-
-        void TestAction(IFormFile? file, TraceIdentifier traceId)
-        {
-            fileArgument = file;
-            traceIdArgument = traceId;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "file", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers[headerName] = headerValue;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-        httpContext.TraceIdentifier = "my-trace-id";
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files["file"], fileArgument);
-        Assert.Equal("file.txt", fileArgument!.FileName);
-        Assert.Equal("file", fileArgument.Name);
-
-        Assert.Equal("my-trace-id", traceIdArgument.Id);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePopulatesFromIFormFileParameterIfRequestHasClientCertificate()
-    {
-        IFormFile? fileArgument = null;
-        TraceIdentifier traceIdArgument = default;
-
-        void TestAction(IFormFile? file, TraceIdentifier traceId)
-        {
-            fileArgument = file;
-            traceIdArgument = traceId;
-        }
-
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "file", "file.txt");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-        httpContext.TraceIdentifier = "my-trace-id";
-
-#pragma warning disable SYSLIB0026 // Type or member is obsolete
-        var clientCertificate = new X509Certificate2();
-#pragma warning restore SYSLIB0026 // Type or member is obsolete
-
-        httpContext.Features.Set<ITlsConnectionFeature>(new TlsConnectionFeature(clientCertificate));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form.Files["file"], fileArgument);
-        Assert.Equal("file.txt", fileArgument!.FileName);
-        Assert.Equal("file", fileArgument.Name);
-
-        Assert.Equal("my-trace-id", traceIdArgument.Id);
     }
 
     public static TheoryData<HttpContent, string> FormContent
@@ -3465,384 +2954,6 @@ public partial class RequestDelegateFactoryTests : LoggedTest
 
         var nse = Assert.Throws<NotSupportedException>(() => RequestDelegateFactory.Create(TestAction));
         Assert.Equal("Assigning a value to the IFromFormMetadata.Name property is not supported for parameters of type IFormCollection.", nse.Message);
-    }
-
-    [Theory]
-    [MemberData(nameof(FormContent))]
-    public async Task RequestDelegatePopulatesFromIFormCollectionParameter(HttpContent content, string contentType)
-    {
-        IFormCollection? formArgument = null;
-
-        void TestAction(IFormCollection formCollection)
-        {
-            formArgument = formCollection;
-        }
-
-        var stream = new MemoryStream();
-        await content.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = contentType;
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form, formArgument);
-        Assert.NotNull(formArgument);
-        Assert.Collection(formArgument!,
-            (item) =>
-            {
-                Assert.Equal("message", item.Key);
-                Assert.Equal("hello", item.Value);
-            },
-            (item) =>
-            {
-                Assert.Equal("name", item.Key);
-                Assert.Equal("foo", item.Value);
-            });
-
-        var allAcceptsMetadata = factoryResult.EndpointMetadata.OfType<IAcceptsMetadata>();
-        var acceptsMetadata = Assert.Single(allAcceptsMetadata);
-
-        Assert.NotNull(acceptsMetadata);
-        Assert.Equal(new[] { "multipart/form-data", "application/x-www-form-urlencoded" }, acceptsMetadata.ContentTypes);
-    }
-
-    [Theory]
-    [MemberData(nameof(FormContent))]
-    public async Task RequestDelegatePopulatesFromIFormCollectionParameterWithAttribute(HttpContent content, string contentType)
-    {
-        IFormCollection? formArgument = null;
-
-        void TestAction([FromForm] IFormCollection formCollection)
-        {
-            formArgument = formCollection;
-        }
-
-        var stream = new MemoryStream();
-        await content.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = contentType;
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form, formArgument);
-        Assert.NotNull(formArgument);
-        Assert.Collection(formArgument!,
-            (item) =>
-            {
-                Assert.Equal("message", item.Key);
-                Assert.Equal("hello", item.Value);
-            },
-            (item) =>
-            {
-                Assert.Equal("name", item.Key);
-                Assert.Equal("foo", item.Value);
-            });
-
-        var allAcceptsMetadata = factoryResult.EndpointMetadata.OfType<IAcceptsMetadata>();
-        var acceptsMetadata = Assert.Single(allAcceptsMetadata);
-
-        Assert.NotNull(acceptsMetadata);
-        Assert.Equal(new[] { "multipart/form-data", "application/x-www-form-urlencoded" }, acceptsMetadata.ContentTypes);
-    }
-
-    [Theory]
-    [MemberData(nameof(FormContent))]
-    public async Task RequestDelegatePopulatesFromOptionalFormParameter(HttpContent content, string contentType)
-    {
-        string? messageArgument = null;
-
-        void TestAction([FromForm] string? message)
-        {
-            messageArgument = message;
-        }
-
-        var stream = new MemoryStream();
-        await content.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = contentType;
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form["message"], messageArgument);
-    }
-
-    [Theory]
-    [MemberData(nameof(FormContent))]
-    public async Task RequestDelegatePopulatesFromMultipleRequiredFormParameters(HttpContent content, string contentType)
-    {
-        string? messageArgument = null;
-        string? nameArgument = null;
-
-        void TestAction([FromForm] string message, [FromForm] string name)
-        {
-            messageArgument = message;
-            nameArgument = name;
-        }
-
-        var stream = new MemoryStream();
-        await content.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = contentType;
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form["message"], messageArgument);
-        Assert.NotNull(messageArgument);
-
-        Assert.Equal(httpContext.Request.Form["name"], nameArgument);
-        Assert.NotNull(nameArgument);
-    }
-
-    [Theory]
-    [MemberData(nameof(FormContent))]
-    public async Task RequestDelegatePopulatesFromOptionalMissingFormParameter(HttpContent content, string contentType)
-    {
-        string? messageArgument = null;
-        string? additionalMessageArgument = null;
-
-        void TestAction([FromForm] string? message, [FromForm] string? additionalMessage)
-        {
-            messageArgument = message;
-            additionalMessageArgument = additionalMessage;
-        }
-
-        var stream = new MemoryStream();
-        await content.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = contentType;
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form["message"], messageArgument);
-        Assert.NotNull(messageArgument);
-        Assert.Null(additionalMessageArgument);
-    }
-
-    [Theory]
-    [MemberData(nameof(FormContent))]
-    public async Task RequestDelegatePopulatesFromFormParameterWithMetadata(HttpContent content, string contentType)
-    {
-        string? textArgument = null;
-
-        void TestAction([FromForm(Name = "message")] string text)
-        {
-            textArgument = text;
-        }
-
-        var stream = new MemoryStream();
-        await content.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = contentType;
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form["message"], textArgument);
-        Assert.NotNull(textArgument);
-    }
-
-    [Theory]
-    [MemberData(nameof(FormContent))]
-    public async Task RequestDelegatePopulatesFromFormAndBoundParameter(HttpContent content, string contentType)
-    {
-        string? messageArgument = null;
-        TraceIdentifier traceIdArgument = default;
-
-        void TestAction([FromForm] string? message, TraceIdentifier traceId)
-        {
-            messageArgument = message;
-            traceIdArgument = traceId;
-        }
-
-        var stream = new MemoryStream();
-        await content.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = contentType;
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-        httpContext.TraceIdentifier = "my-trace-id";
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request.Form["message"], messageArgument);
-        Assert.NotNull(messageArgument);
-
-        Assert.Equal("my-trace-id", traceIdArgument.Id);
-    }
-
-    public static IEnumerable<object[]> FormAndFormFileParametersDelegates
-    {
-        get
-        {
-            void TestAction(HttpContext context, IFormCollection form, IFormFileCollection formFiles)
-            {
-                context.Items["FormFilesArgument"] = formFiles;
-                context.Items["FormArgument"] = form;
-            }
-
-            void TestActionDifferentOrder(HttpContext context, IFormFileCollection formFiles, IFormCollection form)
-            {
-                context.Items["FormFilesArgument"] = formFiles;
-                context.Items["FormArgument"] = form;
-            }
-
-            return new List<object[]>
-                {
-                    new object[] { (Action<HttpContext, IFormCollection, IFormFileCollection>)TestAction },
-                    new object[] { (Action<HttpContext, IFormFileCollection, IFormCollection>)TestActionDifferentOrder },
-                };
-        }
-    }
-
-    [Theory]
-    [MemberData(nameof(FormAndFormFileParametersDelegates))]
-    public async Task RequestDelegatePopulatesFromBothIFormCollectionAndIFormFileParameters(Delegate action)
-    {
-        var fileContent = new StringContent("hello", Encoding.UTF8, "application/octet-stream");
-        var form = new MultipartFormDataContent("some-boundary");
-        form.Add(fileContent, "file", "file.txt");
-        form.Add(new StringContent("foo"), "name");
-
-        var stream = new MemoryStream();
-        await form.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = "multipart/form-data;boundary=some-boundary";
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(action);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        IFormFileCollection? formFilesArgument = httpContext.Items["FormFilesArgument"] as IFormFileCollection;
-        IFormCollection? formArgument = httpContext.Items["FormArgument"] as IFormCollection;
-
-        Assert.Equal(httpContext.Request.Form.Files, formFilesArgument);
-        Assert.NotNull(formFilesArgument!["file"]);
-        Assert.Equal("file.txt", formFilesArgument!["file"]!.FileName);
-
-        Assert.Equal(httpContext.Request.Form, formArgument);
-        Assert.NotNull(formArgument);
-        Assert.Collection(formArgument!,
-            (item) =>
-            {
-                Assert.Equal("name", item.Key);
-                Assert.Equal("foo", item.Value);
-            });
-
-        var allAcceptsMetadata = factoryResult.EndpointMetadata.OfType<IAcceptsMetadata>();
-        Assert.Collection(allAcceptsMetadata,
-            (m) => Assert.Equal(new[] { "multipart/form-data" }, m.ContentTypes));
-    }
-
-    [Theory]
-    [MemberData(nameof(FormContent))]
-    public async Task RequestDelegateSets400ResponseIfRequiredFormItemNotSpecified(HttpContent content, string contentType)
-    {
-        var invoked = false;
-
-        void TestAction([FromForm] string unknownParameter)
-        {
-            invoked = true;
-        }
-
-        var stream = new MemoryStream();
-        await content.CopyToAsync(stream);
-
-        stream.Seek(0, SeekOrigin.Begin);
-
-        var httpContext = CreateHttpContext();
-        httpContext.Request.Body = stream;
-        httpContext.Request.Headers["Content-Type"] = contentType;
-        httpContext.Features.Set<IHttpRequestBodyDetectionFeature>(new RequestBodyDetectionFeature(true));
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.False(invoked);
-        Assert.Equal(400, httpContext.Response.StatusCode);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePopulatesTryParsableParametersFromForm()
-    {
-        var httpContext = CreateHttpContext();
-
-        httpContext.Request.Form = new FormCollection(new Dictionary<string, StringValues>
-        {
-            ["tryParsable"] = "https://example.org"
-        });
-
-        var factoryResult = RequestDelegateFactory.Create((HttpContext httpContext, [FromForm] MyTryParseRecord tryParsable) =>
-        {
-            httpContext.Items["tryParsable"] = tryParsable;
-        });
-
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        var content = Assert.IsType<MyTryParseRecord>(httpContext.Items["tryParsable"]);
-        Assert.Equal(new Uri("https://example.org"), content.Uri);
     }
 
     private record struct ParameterListRecordStruct(HttpContext HttpContext, [FromRoute] int Value);
